Merge https://github.com/nikolas/github-drama/pull/152 into https://github.com/github-drama/github-drama
Merge https://github.com/nikolas/github-drama/pull/152

into https://github.com/github-drama/github-drama

Important Notice:

To edit, open a pull request. We will merge it as soon as we see the notification.

To edit a large amount of content, open an issue saying so. We will grant you write access.

The GitHub drama community forked and created this repository in 2022 because the original maintainer is inactive.

All community members edit collaboratively and enjoy a liberal editing policy similar to Wikipedia.