Summarize the Opal drama by Coraline Ada Ehmke and Meh
Summarize the Opal drama by Coraline Ada Ehmke and Meh.
https://github.com/opal/opal/issues/941
A core maintainer of Opal made transmisic and transphobic comments on the Internet. The creator of the Contributor Covenant opened an issue to kick the maintainer out of the Opal development community. Another core maintainer of Opal replied hostilely and closed the GitHub issue dismissively.