Update project description
Update description to remove promise that every PR is merged as soon as possible. It was left from previous maintainers, but i never intended to actually live up to it. We should have a proper PR policy where only minor PRs that aren't against the project dirction are merged as soon as possible.